Fixes #80 - reopens details
When opening trail details and leaving the map view and returning, the
map details were not being shown. This commit makes it so the
toggleView function checks whether a trail is selected, and if it is
then it displays the trails view, otherwise it displays the maps view.
It also changes DEFAULT_VIEW to MAP_VIEW to better describe what is
being shown when that view is specified.