Fixes #101 - downsizes markers
Downsizes other markers when a trailhead marker is selected.
- Adds a custom CSS class “trail-selected” to map container for when a
trailhead marker is selected, which allows the markers to be customized
with CSS.
- Adds a map zoom level class when zooming the map, so custom CSS can
be added for specific zoom levels, if desired.
- Refactors select/deselect of markers loop so that every marker that
is not selected doesn’t have a call to its deselect() method, but
instead the marker that’s clicked is selected and the last marker
clicked is stored and then deselected when a new marker is clicked.
- Adds the default icon create function for clustering, to set the
stage for creating custom clustering icons.
- Adds functions to move a marker from a marker cluster to the map and
vice versa.
- Removes unnecessary whitespace, where noticed.
- Comment out unused methods in controller (see #103)