Fixes #113 - Filter Tap Performance
Tapping the filters was taking a very long time because the code looped
through all trail segments on each click. This update turns the filters
into a bitmap, so that all four filter values can be stored as a single
integer and bitwise operators can be used to switch them on and off and
check their aggregate value. The values are also stored so they don’t
need to be looked up each click of a filter.